xconsole: update to 1.1.0.
This release fixes use of the scroll wheel in the text area, adds
-version and -help options, and removes support for several older
OS'es.

For those building for 32-bit platforms, it also enables use of the
"large file" APIs - though no file processed by this program should ever
be anywhere near 2GB in size, it is a prerequisite for post-y2038 support.
Since this release was generated using the new GNU autoconf 2.72, this
also adds a --enable-year2038 configure flag which may allow it to
work with files whose timestamps are later than January 19, 2038, but
this has not been tested.

Alan Coopersmith (13):
      configure: Use AC_SYS_LARGEFILE to enable large file support
      Add -version option
      Fix -version handling to not require opening a display first
      Add -help option and usage message when unknown options are given
      man page: resync SYNOPSIS and DESCRIPTION sections
      Allow scrolling in text area with mouse wheel
      unifdef __hpux
      unifdef hpux
      unifdef __UNIXWARE__
      unifdef SCO325
      Remove ifdefs for non-Solaris SysV systems
      unifdef _AIX
      xconsole 1.1.0