devel/py-u-msgpack: import py-u-msgpack-2.8.0
Packaged for wip by K.I.A.Derouiche and myself.

u-msgpack-python is a lightweight serializer and deserializer
module written in pure Python, compatible with both Python 2
and 3, as well CPython and PyPy implementations of Python.
u-msgpack-python is fully compliant with the latest.
In particular, it supports the new binary, UTF-8 string,
application-defined ext, and timestamp types.