py-cheroot: updated to 10.0.1
v10.0.1

Bug fixes

Fixed a flaw where internally unhandled exceptions could crash the worker threads and eventually starve the server of its processing resources. It is no longer and issue and the unhandled errors are now logged and suppressed except for a few expected exceptions that are used for normal interruption requests.

Fixed compatibility with Python 3.8 in the built-in TLS adapter that relies on ssl.
Modern Python versions communicate specialized exceptions ssl.SSLEOFError and ssl.SSLZeroReturnError where the older versions errored out in a very generic way.