pkgtools/url2pkg: use deterministic filesystem lookups
Some filesystems return the filenames in unsorted order, so sort each
result of a 'glob' call. Reported by Adam Ciarcinski.

When looking for candidates for a package in the pkgsrc tree, only
consider directories that contain a Makefile. One of the url2pkg tests
looks for pkglint, and while a 'cvs update' is running, there is an old
entry 'devel/pkglint' in the filesystem that made the test fail.