bsd.prefs.mk: Work around Cygwin OBJECT_FMT mistake in bsd.own.mk.
We should maybe isolate this OBJECT_FMT detection in a way that it
can be written down once, not 2-3 times in bsd.prefs.mk and
bsd.own.mk.

Part of the reason it is duplicated now is because it is done
separately for NATIVE_OBJECT_FMT (the object format used by native
packages we're running) and OBJECT_FMT (the object format used by
packages we're building, which may be cross-built).

The same treatment may be required for AIX, OSF/1, and HP-UX, for
which bsd.prefs.mk sets a non-ELF OBJECT_FMT, but for which
bsd.own.mk appears to incorrectly set OBJECT_FMT=ELF too like it does
for Cygwin and Darwin.

No functional change intended on non-Cygwin platforms because this is
conditional on ${NATIVE_OPSYS} == "Cygwin".

PR pkg/58315: pkgsrc reports "object format ``PE'' which differs from ``ELF''" in cygwin