libnjs: Fix MASTER_SITES.
Unfortunately the confusion between this package and lang/njs being updated to
0.8.4, while this package was configured to still directly fetch the 0.8.2
tarball, has resulted in cdn.netbsd.org containing a broken njs-0.8.4.tar.gz
that is a truncated copy of the 0.8.2 distfile.

Fix variables so this doesn't happen again, but anyone fetching from
cdn.netbsd.org will need to temporarily remove it for these packages until the
file can be purged or updated.